requirements.txt
appdirs==1.4.4
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.4
contourpy==1.1.0
cycler==0.11.0
docker-pycreds==0.4.0
faiss-gpu==1.7.2
filelock==3.12.2
fonttools==4.40.0
FrEIA @ git+https://github.com/VLL-HD/FrEIA.git@a4d3a7db135460e4dd11d4fd7f24b1c97fe7c0d3
gitdb==4.0.10
GitPython==3.1.31
graphviz==0.20.1
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lit==16.0.6
MarkupSafe==2.1.3
matplotlib==3.7.1
mpmath==1.3.0
munch==3.0.0
networkx==3.1
numpy==1.25.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.1
pandas==2.0.2
pathtools==0.1.2
Pillow==9.5.0
pretrainedmodels==0.7.4
protobuf==4.23.3
psutil==5.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
sentry-sdk==1.25.1
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sympy==1.12
termcolor==2.3.0
threadpoolctl==3.1.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
wandb==0.15.4
